backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.3
sqlalchemy==2.0.36
alembic==1.14.0
redis==5.2.0
celery==5.4.0
httpx==0.28.1
python-multipart==0.0.16
cors==1.0.1
aiofiles==24.1.0